Update keyloggergithub.py
This Python script serves as a proof of concept key-logger designed to record keystrokes and transmit them via email. Upon execution, it captures typed keys, storing them in RAM until a predefined limit is reached. At this threshold, the accumulated keystrokes are formatted into an email message and sent, after which the memory storage is cleared to accommodate new keystrokes. The script utilizes threading to concurrently log keystrokes and handle email transmission. The script formats the email for easier readability of the logged keystrokes. If you want you can enhance security by setting an environmental variable to obscure the email password. It's essential to emphasize that this tool is intended solely for educational purposes.